Changed all int and byte types to use stdint instead

+++ include/hashmap.h
@@ -6,19 +6,19 @@
 
 typedef struct {
   void * entries;
-  int count;
-  int capacity;
+  uint32_t count;
+  uint32_t capacity;
   size_t key_size;
   size_t value_size;
-  int (*cmp)(void const * key1, void const  * key2);
-  uint32_t (*hashKey)(void * key, int length);
+  int8_t (*cmp)(void const * key1, void const  * key2);
+  uint32_t (*hashKey)(void * key, size_t length);
 } Map;
 
-void initMap(Map * map, size_t key_size, size_t value_size, int (*cmp)(void const * key1, void const  * key2), uint32_t (*hash)(void * key, int length));
+void initMap(Map * map, size_t key_size, size_t value_size, int8_t (*cmp)(void const * key1, void const  * key2), uint32_t (*hash)(void * key, size_t length));
 void freeMap(Map * map);
 
 //Returns 1 if key-value is a new pair or 0 if key already existed and value was updated
-int mapPut(Map * map, void * key, void * value);
+int8_t mapPut(Map * map, void * key, void * value);
 void mapRemove(Map * map, void * key);
 void * mapGet(Map * map, void * key);
 void mapClear(Map * map);
